+++ components/detail/Rooms/Index.tsx
@@ -3,7 +3,8 @@ import React from 'react';
 import { roomList } from '@/data/detail/roomData';
 import { useRouter } from 'next/router';
 import { formatNumberWithCommas } from 'utils/formatNumberWithCommas';
-import { BlackButton, WhiteButton } from '../Top/global/button';
+import { WhiteButton } from '@/components/global/button/WhiteButton';
+import { BlackButton } from '@/components/global/button/BlackButton';
 
 type Props = {};
