+++ components/detail/Top/Index.tsx
@@ -8,10 +8,10 @@ import heartFillIcon from '@/public/images/heart_fill.svg';
 import { hotelInfo } from '@/data/detail/hotelData';
 import { formatNumberWithCommas } from 'utils/formatNumberWithCommas';
 import { motion } from 'framer-motion';
-import { BlackButton } from './global/button';
 import { fetchProductData } from 'api/reviewApi';
 import { productId } from '@/data/detail/productData';
 import { ProductDataType } from 'api/types/reviewTypes';
+import { BlackButton } from '@/components/global/button/BlackButton';
 
 export default function TopInfo() {
   const [productData, setProductData] = useState<ProductDataType | null>(null);
